Web API: /sdrangel/preset (PATCH) implementation

+++ swagger/sdrangel/code/qt5/client/SWGPresetIdentifier.h
@@ -45,8 +45,8 @@ public:
     QString* getGroupName();
     void setGroupName(QString* group_name);
 
-    float getCenterFrequency();
-    void setCenterFrequency(float center_frequency);
+    qint64 getCenterFrequency();
+    void setCenterFrequency(qint64 center_frequency);
 
     QString* getType();
     void setType(QString* type);
@@ -57,7 +57,7 @@ public:
 
 private:
     QString* group_name;
-    float center_frequency;
+    qint64 center_frequency;
     QString* type;
     QString* name;
 };
